refactor(billing): make stripe portal single source of truth for billing settings
Replace editable account and org billing forms with read-only Stripe customer data plus a portal link, and expose customer email on billing-details reads.
