+++ apps/core/src/routes/v1/organizations/[id]/billing-details/get.ts
@@ -31,6 +31,7 @@ const route = createRoute({
       {
         data: {
           stripeCustomerId: "cus_org_123",
+          email: "billing@acme.example",
           address: null,
           taxIds: [],
         },
