+++ apps/core/src/routes/v1/users/[id]/billing-details/get.ts
@@ -35,6 +35,7 @@ const route = withGlobalHeaderParameters(
         {
           data: {
             stripeCustomerId: "cus_123",
+            email: "billing@example.com",
             address: {
               line1: "123 Main St",
               line2: null,
