+++ apps/web/messages/de.json
@@ -2938,8 +2938,8 @@
         "credits": "Credits",
         "coupon": "Gutschein"
       },
-      "billingPortalDescription": "Lade deine Rechnungen herunter, ändere deine Rechnungsadressen und verwalte weitere Rechnungsinformationen.",
-      "manageYourBilling": "Billing verwalten",
+      "billingPortalDescription": "Rechnungsadresse, Steuer-ID, Zahlungsmethoden, Rechnungs-E-Mail und Abonnements in Stripe verwalten.",
+      "manageYourBilling": "In Stripe verwalten",
       "openingBillingPortal": "Wird geöffnet...",
       "orgAccessRestrictedTitle": "Billing-Zugang eingeschränkt",
       "orgAccessRestrictedDescription": "Nur Eigentümer und Admins der Organisation können Billing-Details einsehen.",
@@ -3168,33 +3168,18 @@
         "error": "Benachrichtigungseinstellung konnte nicht aktualisiert werden"
       },
       "BillingDetails": {
-        "title": "Rechnungsadresse",
-        "description": "Adresse und Steuer-ID für Rechnungen und Steuerberechnung. Gespeichert in deinem Stripe-Kundenprofil.",
-        "line1Label": "Adresszeile 1",
-        "line2Label": "Adresszeile 2 (optional)",
-        "cityLabel": "Stadt",
-        "stateLabel": "Bundesland / Provinz",
-        "postalCodeLabel": "Postleitzahl",
-        "countryLabel": "Land",
-        "countryPlaceholder": "Land auswählen",
+        "title": "Rechnungsinformationen",
+        "description": "Rechnungsadresse, Steuer-ID, Rechnungs-E-Mail und Zahlungsmethoden werden in Stripe verwaltet.",
+        "addressLabel": "Rechnungsadresse",
         "taxIdLabel": "USt-IdNr. / Steuer-ID",
-        "taxIdPlaceholder": "z. B. DE123456789",
-        "taxIdHelp": "Optional. Wird für die steuerliche Behandlung von Unternehmen verwendet, z. B. EU-Reverse-Charge.",
         "taxIdVerification": "Verifizierungsstatus: {status}",
-        "empty": "Noch keine Rechnungsadresse hinterlegt.",
-        "add": "Rechnungsadresse hinzufügen",
-        "edit": "Bearbeiten",
-        "save": "Speichern",
-        "cancel": "Abbrechen",
-        "Success": {
-          "update": "Rechnungsdetails erfolgreich aktualisiert"
-        },
+        "invoiceEmailLabel": "Rechnungs-E-Mail",
+        "invoiceEmailEmpty": "Keine Rechnungs-E-Mail hinterlegt",
+        "empty": "Keine Rechnungsadresse hinterlegt",
         "Errors": {
           "unauthenticated": "Bitte melde dich an, um fortzufahren",
           "unauthenticatedAction": "Anmelden",
-          "unauthorized": "Du hast keine Berechtigung, Rechnungsdetails zu aktualisieren",
-          "badInput": "Bitte überprüfe Rechnungsadresse und Steuer-ID",
-          "general": "Rechnungsdetails konnten nicht aktualisiert werden"
+          "general": "Stripe konnte nicht geöffnet werden"
         }
       },
       "Delete": {
@@ -3438,54 +3423,20 @@
         "delete": "Löschen",
         "invite": "Mitglied einladen",
         "bulkInvite": "Mehrere einladen",
-        "InvoiceEmail": {
-          "title": "Rechnungs-E-Mail",
-          "description": "E-Mail-Adresse für den Empfang von Rechnungen und Billing-Benachrichtigungen",
-          "label": "Rechnungs-E-Mail",
-          "placeholder": "billing@example.com",
-          "help": "Diese E-Mail-Adresse erhält alle Rechnungen und billing-bezogenen Mitteilungen von Stripe",
-          "edit": "Bearbeiten",
-          "save": "Speichern",
-          "cancel": "Abbrechen",
-          "Success": {
-            "update": "Rechnungs-E-Mail erfolgreich aktualisiert"
-          },
-          "Errors": {
-            "unauthenticated": "Bitte melde dich an, um fortzufahren",
-            "unauthenticatedAction": "Anmelden",
-            "unauthorized": "Nur Eigentümer und Admins der Organisation können die Rechnungs-E-Mail aktualisieren",
-            "badInput": "Bitte gib eine gültige E-Mail-Adresse ein",
-            "general": "Rechnungs-E-Mail konnte nicht aktualisiert werden"
-          }
-        },
         "BillingDetails": {
-          "title": "Rechnungsadresse",
-          "description": "Adresse und Steuer-ID für Organisationsrechnungen und Steuerberechnung. Gespeichert im Stripe-Kundenprofil der Organisation.",
-          "line1Label": "Adresszeile 1",
-          "line2Label": "Adresszeile 2 (optional)",
-          "cityLabel": "Stadt",
-          "stateLabel": "Bundesland / Provinz",
-          "postalCodeLabel": "Postleitzahl",
-          "countryLabel": "Land",
-          "countryPlaceholder": "Land auswählen",
+          "title": "Rechnungsinformationen",
+          "description": "Rechnungsadresse, Steuer-ID, Rechnungs-E-Mail und Zahlungsmethoden werden in Stripe verwaltet.",
+          "addressLabel": "Rechnungsadresse",
           "taxIdLabel": "USt-IdNr. / Steuer-ID",
-          "taxIdPlaceholder": "z. B. DE123456789",
-          "taxIdHelp": "Optional. Wird für die steuerliche Behandlung von Unternehmen verwendet, z. B. EU-Reverse-Charge.",
           "taxIdVerification": "Verifizierungsstatus: {status}",
-          "empty": "Noch keine Rechnungsadresse hinterlegt.",
-          "add": "Rechnungsadresse hinzufügen",
-          "edit": "Bearbeiten",
-          "save": "Speichern",
-          "cancel": "Abbrechen",
-          "Success": {
-            "update": "Rechnungsdetails erfolgreich aktualisiert"
-          },
+          "invoiceEmailLabel": "Rechnungs-E-Mail",
+          "invoiceEmailEmpty": "Keine Rechnungs-E-Mail hinterlegt",
+          "empty": "Keine Rechnungsadresse hinterlegt",
           "Errors": {
             "unauthenticated": "Bitte melde dich an, um fortzufahren",
             "unauthenticatedAction": "Anmelden",
-            "unauthorized": "Nur Eigentümer und Admins der Organisation können Rechnungsdetails aktualisieren",
-            "badInput": "Bitte überprüfe Rechnungsadresse und Steuer-ID",
-            "general": "Rechnungsdetails konnten nicht aktualisiert werden"
+            "unauthorized": "Nur Eigentümer und Admins der Organisation können Stripe-Billing öffnen",
+            "general": "Stripe konnte nicht geöffnet werden"
           }
         },
         "Subscription": {
