+++ apps/web/messages/fr.json
@@ -2969,8 +2969,8 @@
         "credits": "Crédits",
         "coupon": "Code promo"
       },
-      "billingPortalDescription": "Téléchargez vos factures, modifiez vos adresses de facturation et gérez d’autres informations de facturation.",
-      "manageYourBilling": "Gérer votre facturation",
+      "billingPortalDescription": "Mettez à jour l'adresse de facturation, l'identifiant fiscal, les moyens de paiement, l'e-mail de facturation et les abonnements dans Stripe.",
+      "manageYourBilling": "Gérer dans Stripe",
       "openingBillingPortal": "Ouverture...",
       "orgAccessRestrictedTitle": "Accès à la facturation restreint",
       "orgAccessRestrictedDescription": "Seuls les propriétaires et administrateurs de l’organisation peuvent consulter les informations de facturation de cette organisation.",
@@ -3199,33 +3199,18 @@
         "error": "Échec de la mise à jour des préférences de notification"
       },
       "BillingDetails": {
-        "title": "Adresse de facturation",
-        "description": "Adresse et numéro de TVA utilisés pour les factures et le calcul des taxes. Stockés sur votre profil client Stripe.",
-        "line1Label": "Adresse ligne 1",
-        "line2Label": "Adresse ligne 2 (facultatif)",
-        "cityLabel": "Ville",
-        "stateLabel": "État / province",
-        "postalCodeLabel": "Code postal",
-        "countryLabel": "Pays",
-        "countryPlaceholder": "Sélectionner un pays",
+        "title": "Informations de facturation",
+        "description": "L'adresse de facturation, l'identifiant fiscal, l'e-mail de facturation et les moyens de paiement sont gérés dans Stripe.",
+        "addressLabel": "Adresse de facturation",
         "taxIdLabel": "TVA / identifiant fiscal",
-        "taxIdPlaceholder": "ex. DE123456789",
-        "taxIdHelp": "Facultatif. Utilisé pour le traitement fiscal des entreprises, par ex. l'autoliquidation UE.",
         "taxIdVerification": "Statut de vérification : {status}",
-        "empty": "Aucune adresse de facturation enregistrée.",
-        "add": "Ajouter une adresse de facturation",
-        "edit": "Modifier",
-        "save": "Enregistrer",
-        "cancel": "Annuler",
-        "Success": {
-          "update": "Informations de facturation mises à jour"
-        },
+        "invoiceEmailLabel": "E-mail de facturation",
+        "invoiceEmailEmpty": "Aucun e-mail de facturation enregistré",
+        "empty": "Aucune adresse de facturation enregistrée",
         "Errors": {
           "unauthenticated": "Veuillez vous connecter pour continuer",
           "unauthenticatedAction": "Se connecter",
-          "unauthorized": "Vous n'avez pas l'autorisation de modifier les informations de facturation",
-          "badInput": "Vérifiez l'adresse de facturation et l'identifiant fiscal",
-          "general": "Échec de la mise à jour des informations de facturation"
+          "general": "Impossible d'ouvrir Stripe"
         }
       },
       "Delete": {
@@ -3469,54 +3454,20 @@
         "delete": "Supprimer",
         "invite": "Inviter un membre",
         "bulkInvite": "Invitation groupée",
-        "InvoiceEmail": {
-          "title": "E-mail de facturation",
-          "description": "Adresse e-mail pour recevoir les factures et les notifications de facturation",
-          "label": "E-mail de facturation",
-          "placeholder": "billing@example.com",
-          "help": "Cet e-mail recevra toutes les factures et communications liées à la facturation de Stripe",
-          "edit": "Modifier",
-          "save": "Enregistrer",
-          "cancel": "Annuler",
-          "Success": {
-            "update": "E-mail de facturation mis à jour avec succès"
-          },
-          "Errors": {
-            "unauthenticated": "Veuillez vous connecter pour continuer",
-            "unauthenticatedAction": "Se connecter",
-            "unauthorized": "Seuls les propriétaires et administrateurs de l’organisation peuvent mettre à jour l’e-mail de facturation",
-            "badInput": "Veuillez saisir une adresse e-mail valide",
-            "general": "Échec de la mise à jour de l’e-mail de facturation"
-          }
-        },
         "BillingDetails": {
-          "title": "Adresse de facturation",
-          "description": "Adresse et numéro de TVA utilisés pour les factures de l'organisation et le calcul des taxes. Stockés sur le profil client Stripe de l'organisation.",
-          "line1Label": "Adresse ligne 1",
-          "line2Label": "Adresse ligne 2 (facultatif)",
-          "cityLabel": "Ville",
-          "stateLabel": "État / province",
-          "postalCodeLabel": "Code postal",
-          "countryLabel": "Pays",
-          "countryPlaceholder": "Sélectionner un pays",
+          "title": "Informations de facturation",
+          "description": "L'adresse de facturation, l'identifiant fiscal, l'e-mail de facturation et les moyens de paiement sont gérés dans Stripe.",
+          "addressLabel": "Adresse de facturation",
           "taxIdLabel": "TVA / identifiant fiscal",
-          "taxIdPlaceholder": "ex. DE123456789",
-          "taxIdHelp": "Facultatif. Utilisé pour le traitement fiscal des entreprises, par ex. l'autoliquidation UE.",
           "taxIdVerification": "Statut de vérification : {status}",
-          "empty": "Aucune adresse de facturation enregistrée.",
-          "add": "Ajouter une adresse de facturation",
-          "edit": "Modifier",
-          "save": "Enregistrer",
-          "cancel": "Annuler",
-          "Success": {
-            "update": "Informations de facturation mises à jour"
-          },
+          "invoiceEmailLabel": "E-mail de facturation",
+          "invoiceEmailEmpty": "Aucun e-mail de facturation enregistré",
+          "empty": "Aucune adresse de facturation enregistrée",
           "Errors": {
             "unauthenticated": "Veuillez vous connecter pour continuer",
             "unauthenticatedAction": "Se connecter",
-            "unauthorized": "Seuls les propriétaires et administrateurs de l'organisation peuvent modifier les informations de facturation",
-            "badInput": "Vérifiez l'adresse de facturation et l'identifiant fiscal",
-            "general": "Échec de la mise à jour des informations de facturation"
+            "unauthorized": "Seuls les propriétaires et administrateurs peuvent ouvrir la facturation Stripe",
+            "general": "Impossible d'ouvrir Stripe"
           }
         },
         "Subscription": {
