+++ apps/web/messages/ja.json
@@ -2969,8 +2969,8 @@
         "credits": "クレジット",
         "coupon": "クーポン"
       },
-      "billingPortalDescription": "請求書をダウンロードし、請求先住所を変更し、その他の請求情報を管理できます。",
-      "manageYourBilling": "請求を管理",
+      "billingPortalDescription": "Stripeで請求先住所、税ID、支払い方法、請求書メール、サブスクリプションを管理します。",
+      "manageYourBilling": "Stripe で管理",
       "openingBillingPortal": "開いています...",
       "orgAccessRestrictedTitle": "請求へのアクセスが制限されています",
       "orgAccessRestrictedDescription": "この組織の請求情報を閲覧できるのは、組織のオーナーと管理者のみです。",
@@ -3199,33 +3199,18 @@
         "error": "通知設定の更新に失敗しました"
       },
       "BillingDetails": {
-        "title": "請求先住所",
-        "description": "請求書と税金計算に使用する住所と税 ID。Stripe 顧客プロフィールに保存されます。",
-        "line1Label": "住所 1",
-        "line2Label": "住所 2（任意）",
-        "cityLabel": "市区町村",
-        "stateLabel": "都道府県 / 州",
-        "postalCodeLabel": "郵便番号",
-        "countryLabel": "国",
-        "countryPlaceholder": "国を選択",
+        "title": "請求情報",
+        "description": "請求先住所、税 ID、請求書メール、支払い方法は Stripe で管理されます。",
+        "addressLabel": "請求先住所",
         "taxIdLabel": "VAT / 税 ID",
-        "taxIdPlaceholder": "例: DE123456789",
-        "taxIdHelp": "任意。EU リバースチャージなど、法人の税務処理に使用されます。",
         "taxIdVerification": "確認ステータス: {status}",
-        "empty": "請求先住所はまだ登録されていません。",
-        "add": "請求先住所を追加",
-        "edit": "編集",
-        "save": "保存",
-        "cancel": "キャンセル",
-        "Success": {
-          "update": "請求情報を更新しました"
-        },
+        "invoiceEmailLabel": "請求書送付先メール",
+        "invoiceEmailEmpty": "請求書送付先メールは未登録です",
+        "empty": "請求先住所は未登録です",
         "Errors": {
           "unauthenticated": "続行するにはログインしてください",
           "unauthenticatedAction": "ログイン",
-          "unauthorized": "請求情報を更新する権限がありません",
-          "badInput": "請求先住所と税 ID を確認してください",
-          "general": "請求情報の更新に失敗しました"
+          "general": "Stripe を開けませんでした"
         }
       },
       "Delete": {
@@ -3469,54 +3454,20 @@
         "delete": "削除",
         "invite": "メンバーを招待",
         "bulkInvite": "一括招待",
-        "InvoiceEmail": {
-          "title": "請求書送付先メール",
-          "description": "請求書および請求関連の通知を受け取るメールアドレス",
-          "label": "請求書送付先メール",
-          "placeholder": "billing@example.com",
-          "help": "このメールアドレス宛てに、Stripe からの請求書および請求に関する連絡が届きます",
-          "edit": "編集",
-          "save": "保存",
-          "cancel": "キャンセル",
-          "Success": {
-            "update": "請求書送付先メールを更新しました"
-          },
-          "Errors": {
-            "unauthenticated": "続行するにはログインしてください",
-            "unauthenticatedAction": "ログイン",
-            "unauthorized": "請求書送付先メールを更新できるのは、組織のオーナーと管理者のみです",
-            "badInput": "有効なメールアドレスを入力してください",
-            "general": "請求書送付先メールの更新に失敗しました"
-          }
-        },
         "BillingDetails": {
-          "title": "請求先住所",
-          "description": "組織の請求書と税金計算に使用する住所と税 ID。組織の Stripe 顧客プロフィールに保存されます。",
-          "line1Label": "住所 1",
-          "line2Label": "住所 2（任意）",
-          "cityLabel": "市区町村",
-          "stateLabel": "都道府県 / 州",
-          "postalCodeLabel": "郵便番号",
-          "countryLabel": "国",
-          "countryPlaceholder": "国を選択",
+          "title": "請求情報",
+          "description": "請求先住所、税 ID、請求書メール、支払い方法は Stripe で管理されます。",
+          "addressLabel": "請求先住所",
           "taxIdLabel": "VAT / 税 ID",
-          "taxIdPlaceholder": "例: DE123456789",
-          "taxIdHelp": "任意。EU リバースチャージなど、法人の税務処理に使用されます。",
           "taxIdVerification": "確認ステータス: {status}",
-          "empty": "請求先住所はまだ登録されていません。",
-          "add": "請求先住所を追加",
-          "edit": "編集",
-          "save": "保存",
-          "cancel": "キャンセル",
-          "Success": {
-            "update": "請求情報を更新しました"
-          },
+          "invoiceEmailLabel": "請求書送付先メール",
+          "invoiceEmailEmpty": "請求書送付先メールは未登録です",
+          "empty": "請求先住所は未登録です",
           "Errors": {
             "unauthenticated": "続行するにはログインしてください",
             "unauthenticatedAction": "ログイン",
-            "unauthorized": "組織のオーナーまたは管理者のみ請求情報を更新できます",
-            "badInput": "請求先住所と税 ID を確認してください",
-            "general": "請求情報の更新に失敗しました"
+            "unauthorized": "Stripe の請求管理を開けるのはオーナーと管理者のみです",
+            "general": "Stripe を開けませんでした"
           }
         },
         "Subscription": {
