+++ apps/web/messages/zh-Hans.json
@@ -2969,8 +2969,8 @@
         "credits": "积分",
         "coupon": "优惠券"
       },
-      "billingPortalDescription": "下载您的发票，更改账单地址并管理其他账单信息。",
-      "manageYourBilling": "管理账单",
+      "billingPortalDescription": "在 Stripe 中更新账单地址、税号、支付方式、发票邮箱和订阅。",
+      "manageYourBilling": "在 Stripe 中管理",
       "openingBillingPortal": "正在打开...",
       "orgAccessRestrictedTitle": "计费访问受限",
       "orgAccessRestrictedDescription": "只有组织所有者和管理员可以查看该组织的计费详情。",
@@ -3199,33 +3199,18 @@
         "error": "更新通知偏好失败"
       },
       "BillingDetails": {
-        "title": "账单地址",
-        "description": "用于发票和税费计算的地址与税号。保存在你的 Stripe 客户资料中。",
-        "line1Label": "地址行 1",
-        "line2Label": "地址行 2（可选）",
-        "cityLabel": "城市",
-        "stateLabel": "省 / 州",
-        "postalCodeLabel": "邮政编码",
-        "countryLabel": "国家/地区",
-        "countryPlaceholder": "选择国家/地区",
+        "title": "账单信息",
+        "description": "账单地址、税号、发票邮箱和支付方式均在 Stripe 中管理。",
+        "addressLabel": "账单地址",
         "taxIdLabel": "增值税 / 税号",
-        "taxIdPlaceholder": "例如 DE123456789",
-        "taxIdHelp": "可选。用于企业税务处理，例如欧盟反向征收。",
         "taxIdVerification": "验证状态：{status}",
-        "empty": "尚未保存账单地址。",
-        "add": "添加账单地址",
-        "edit": "编辑",
-        "save": "保存",
-        "cancel": "取消",
-        "Success": {
-          "update": "账单信息更新成功"
-        },
+        "invoiceEmailLabel": "发票邮箱",
+        "invoiceEmailEmpty": "尚未设置发票邮箱",
+        "empty": "尚未保存账单地址",
         "Errors": {
           "unauthenticated": "请登录后继续",
           "unauthenticatedAction": "登录",
-          "unauthorized": "你没有权限更新账单信息",
-          "badInput": "请检查账单地址和税号",
-          "general": "更新账单信息失败"
+          "general": "无法打开 Stripe"
         }
       },
       "Delete": {
@@ -3469,54 +3454,20 @@
         "delete": "删除",
         "invite": "邀请成员",
         "bulkInvite": "批量邀请",
-        "InvoiceEmail": {
-          "title": "发票邮箱",
-          "description": "用于接收发票和计费通知的邮箱地址",
-          "label": "发票邮箱",
-          "placeholder": "billing@example.com",
-          "help": "该邮箱将接收来自 Stripe 的所有发票和计费相关通信",
-          "edit": "编辑",
-          "save": "保存",
-          "cancel": "取消",
-          "Success": {
-            "update": "发票邮箱更新成功"
-          },
-          "Errors": {
-            "unauthenticated": "请登录以继续",
-            "unauthenticatedAction": "登录",
-            "unauthorized": "只有组织所有者和管理员可以更新发票邮箱",
-            "badInput": "请输入有效的邮箱地址",
-            "general": "更新发票邮箱失败"
-          }
-        },
         "BillingDetails": {
-          "title": "账单地址",
-          "description": "用于组织发票和税费计算的地址与税号。保存在组织的 Stripe 客户资料中。",
-          "line1Label": "地址行 1",
-          "line2Label": "地址行 2（可选）",
-          "cityLabel": "城市",
-          "stateLabel": "省 / 州",
-          "postalCodeLabel": "邮政编码",
-          "countryLabel": "国家/地区",
-          "countryPlaceholder": "选择国家/地区",
+          "title": "账单信息",
+          "description": "账单地址、税号、发票邮箱和支付方式均在 Stripe 中管理。",
+          "addressLabel": "账单地址",
           "taxIdLabel": "增值税 / 税号",
-          "taxIdPlaceholder": "例如 DE123456789",
-          "taxIdHelp": "可选。用于企业税务处理，例如欧盟反向征收。",
           "taxIdVerification": "验证状态：{status}",
-          "empty": "尚未保存账单地址。",
-          "add": "添加账单地址",
-          "edit": "编辑",
-          "save": "保存",
-          "cancel": "取消",
-          "Success": {
-            "update": "账单信息更新成功"
-          },
+          "invoiceEmailLabel": "发票邮箱",
+          "invoiceEmailEmpty": "尚未设置发票邮箱",
+          "empty": "尚未保存账单地址",
           "Errors": {
             "unauthenticated": "请登录后继续",
             "unauthenticatedAction": "登录",
-            "unauthorized": "仅组织所有者和管理员可以更新账单信息",
-            "badInput": "请检查账单地址和税号",
-            "general": "更新账单信息失败"
+            "unauthorized": "仅组织所有者和管理员可以打开 Stripe 账单管理",
+            "general": "无法打开 Stripe"
           }
         },
         "Subscription": {
