+++ apps/web/src/app/(app)/account/components/account-billing-details.tsx
@@ -1,7 +1,9 @@
 "use client";
 
-import { BillingDetailsSection } from "@/components/billing/billing-details-section";
-import { updateMyBillingDetails } from "@/lib/actions";
+import { useTranslations } from "next-intl";
+
+import { BalanceBillingPortalLink } from "@/components/billing/balance-billing-portal-link";
+import { StripeBillingInformationCard } from "@/components/billing/stripe-billing-information-card";
 import type { StripeCustomerBillingDetails } from "@/lib/clients/generated/core";
 
 interface AccountBillingDetailsProps {
@@ -11,29 +13,26 @@ interface AccountBillingDetailsProps {
 export function AccountBillingDetails({
   billingDetails,
 }: AccountBillingDetailsProps) {
+  const t = useTranslations("App.Account.BillingDetails");
+  const tBilling = useTranslations("App.Billing");
+
+  const portalLink = billingDetails.stripeCustomerId ? (
+    <BalanceBillingPortalLink
+      description={tBilling("billingPortalDescription")}
+      generalErrorMessage={t("Errors.general")}
+      label={tBilling("manageYourBilling")}
+      openingLabel={tBilling("openingBillingPortal")}
+      returnPath="/account"
+      unauthenticatedActionLabel={t("Errors.unauthenticatedAction")}
+      unauthenticatedErrorMessage={t("Errors.unauthenticated")}
+    />
+  ) : null;
+
   return (
-    <BillingDetailsSection
+    <StripeBillingInformationCard
       billingDetails={billingDetails}
-      canEdit
+      portalLink={portalLink}
       translationNamespace="App.Account.BillingDetails"
-      onSave={async (data) => {
-        const result = await updateMyBillingDetails({
-          address: data.address,
-          taxIdValue: data.taxIdValue,
-        });
-
-        if (result.ok) {
-          return { ok: true as const };
-        }
-
-        return {
-          ok: false as const,
-          error: {
-            code: result.error.code,
-            message: result.error.message,
-          },
-        };
-      }}
     />
   );
 }
