--- apps/web/src/lib/actions/billing/index.ts
@@ -1,1 +1,0 @@
-export * from "./action";
