+++ apps/web/src/lib/actions/index.ts
@@ -1,5 +1,4 @@
 export * from "./auth";
-export * from "./billing";
 export * from "./conversation";
 export * from "./credits";
 export * from "./design-md";
